DecepitNet/requirements.txt
# ============================================================================
# PRODUCTION REQUIREMENTS - DECEPTINET (Sprints 0-9)
# ============================================================================

# Core Dependencies
numpy>=1.24.3
pandas>=2.1.1
scikit-learn>=1.3.1

# Deep Learning
tensorflow>=2.14.0
keras>=2.14.0
torch>=2.1.0
torchvision>=0.16.0

# XGBoost & Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.1.0

# Network & Security
scapy>=2.5.0
paramiko>=3.3.0
cryptography>=41.0.5
pyOpenSSL>=23.2.0
bcrypt>=4.0.1
pyjwt>=2.8.0
python-jose>=3.3.0
passlib>=1.7.4
hvac>=1.2.1  # HashiCorp Vault

# Web & API
flask>=3.0.0
flask-socketio>=5.3.4
flask-cors>=4.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
httpx>=0.25.0
websockets>=12.0
python-multipart>=0.0.6
requests>=2.31.0
aiohttp>=3.8.5

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
redis>=5.0.0
pymongo>=4.5.0
motor>=3.3.1  # Async MongoDB
elasticsearch>=8.10.0

# Message Brokers
aiokafka>=0.8.1
kafka-python>=2.0.2
pika>=1.3.2  # RabbitMQ
celery>=5.3.4

# Monitoring & Logging
prometheus-client>=0.18.0
python-json-logger>=2.0.7
colorlog>=6.7.0
psutil>=5.9.5
py-cpuinfo>=9.0.0

# ML & AI - Reinforcement Learning
ray[rllib]>=2.7.0
gymnasium>=0.29.1
stable-baselines3>=2.1.0
sb3-contrib>=2.1.0
gym>=0.26.2
transformers>=4.35.0

# Explainability (XAI)
shap>=0.43.0
lime>=0.2.0.1

# Threat Intelligence
pymisp>=2.4.175
stix2>=3.0.1
taxii2-client>=2.3.0
mitreattack-python>=3.0.0

# Honeypot Protocols
pyftpdlib>=1.5.7  # FTP
twisted>=23.8.0  # Multiple protocols

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.1
click>=8.1.7
rich>=13.6.0
tabulate>=0.9.0
python-dateutil>=2.8.2
pytz>=2023.3
jsonschema>=4.19.1
marshmallow>=3.20.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
dnspython>=2.4.2

# Network Analysis
python-nmap>=0.7.1
pyshark>=0.6

# Network Simulation
faker>=19.12.0

# Async & Concurrency
asyncio>=3.4.3
aiofiles>=23.2.1

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Deployment & Orchestration
kubernetes>=28.1.0
docker>=6.1.3
ansible>=8.5.0

# ML Optimization
optuna>=3.4.0  # Hyperparameter optimization
mlflow>=2.7.1  # Experiment tracking

# Testing & Quality
pytest>=7.4.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.11.1
black>=23.9.1
flake8>=6.1.0
mypy>=1.5.1
isort>=5.12.0
pylint>=3.0.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.2
sphinx>=7.2.6
